added settings header
